Started extracting JITM related logic to the reusable class
