+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/jitm/JitmStateProvider.kt
@@ -1,0 +1,141 @@
+package com.woocommerce.android.ui.jitm
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import com.woocommerce.android.R
+import com.woocommerce.android.di.AppCoroutineScope
+import com.woocommerce.android.model.UiString
+import com.woocommerce.android.tools.SelectedSite
+import com.woocommerce.android.ui.mystore.MyStoreUtmProvider
+import com.woocommerce.android.ui.mystore.MyStoreViewModel
+import com.woocommerce.android.ui.payments.banner.BannerState
+import com.woocommerce.android.util.WooLog
+import com.woocommerce.android.viewmodel.MultiLiveEvent
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
+import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooResult
+import org.wordpress.android.fluxc.network.rest.wpcom.wc.jitm.JITMApiResponse
+import org.wordpress.android.fluxc.store.JitmStore
+import javax.inject.Inject
+
+interface JitmStateProvider {
+    val bannerState: LiveData<BannerState>
+    val jitmCtaClickedEvent: MutableLiveData<CtaClick>
+
+    data class CtaClick(val url: String) : MultiLiveEvent.Event()
+}
+
+class JitmStateProviderImpl @Inject constructor(
+    private val jitmStore: JitmStore,
+    private val jitmTracker: JitmTracker,
+    private val myStoreUtmProvider: MyStoreUtmProvider,
+    private val queryParamsEncoder: QueryParamsEncoder,
+    private val selectedSite: SelectedSite,
+    @AppCoroutineScope private val appCoroutineScope: CoroutineScope,
+) : JitmStateProvider {
+    private val _bannerState: MutableLiveData<BannerState> = MutableLiveData()
+    override val bannerState: LiveData<BannerState> = _bannerState
+
+    override val jitmCtaClickedEvent: MutableLiveData<JitmStateProvider.CtaClick> = MultiLiveEvent()
+
+    suspend fun fetchJitms(jitmMessagePath: String) {
+        val response = jitmStore.fetchJitmMessage(
+            selectedSite.get(),
+            jitmMessagePath,
+            queryParamsEncoder.getEncodedQueryParams(),
+        )
+        populateResultToUI(response, jitmMessagePath)
+    }
+
+    private fun populateResultToUI(response: WooResult<Array<JITMApiResponse>>, jitmMessagePath: String) {
+        if (response.isError) {
+            jitmTracker.trackJitmFetchFailure(MyStoreViewModel.UTM_SOURCE, response.error.type, response.error.message)
+            WooLog.e(WooLog.T.JITM, "Failed to fetch JITM for the message path $jitmMessagePath")
+            return
+        }
+
+        jitmTracker.trackJitmFetchSuccess(
+            MyStoreViewModel.UTM_SOURCE,
+            response.model?.getOrNull(0)?.id,
+            response.model?.size
+        )
+        response.model?.getOrNull(0)?.let { model: JITMApiResponse ->
+            jitmTracker.trackJitmDisplayed(
+                MyStoreViewModel.UTM_SOURCE,
+                model.id,
+                model.featureClass
+            )
+            _bannerState.value = BannerState.DisplayBannerState(
+                onPrimaryActionClicked = {
+                    onJitmCtaClicked(
+                        id = model.id,
+                        featureClass = model.featureClass,
+                        url = model.cta.link
+                    )
+                },
+                onDismissClicked = {
+                    onJitmDismissClicked(
+                        model.id,
+                        model.featureClass
+                    )
+                },
+                title = UiString.UiStringText(model.content.message),
+                description = UiString.UiStringText(model.content.description),
+                primaryActionLabel = UiString.UiStringText(model.cta.message),
+                chipLabel = UiString.UiStringRes(R.string.card_reader_upsell_card_reader_banner_new)
+            )
+        } ?: run {
+            _bannerState.value = BannerState.HideBannerState
+            WooLog.i(WooLog.T.JITM, "No JITM Campaign in progress")
+        }
+    }
+
+    private fun onJitmCtaClicked(
+        id: String,
+        featureClass: String,
+        url: String
+    ) {
+        jitmTracker.trackJitmCtaTapped(
+            MyStoreViewModel.UTM_SOURCE,
+            id,
+            featureClass
+        )
+        jitmCtaClickedEvent.value =
+            JitmStateProvider.CtaClick(
+                myStoreUtmProvider.getUrlWithUtmParams(
+                    source = MyStoreViewModel.UTM_SOURCE,
+                    id = id,
+                    featureClass = featureClass,
+                    siteId = selectedSite.getIfExists()?.siteId,
+                    url = url
+                )
+            )
+    }
+
+    private fun onJitmDismissClicked(jitmId: String, featureClass: String) {
+        _bannerState.value = BannerState.HideBannerState
+        jitmTracker.trackJitmDismissTapped(MyStoreViewModel.UTM_SOURCE, jitmId, featureClass)
+        appCoroutineScope.launch {
+            jitmStore.dismissJitmMessage(selectedSite.get(), jitmId, featureClass).also { response ->
+                when {
+                    response.model != null && response.model!! -> {
+                        jitmTracker.trackJitmDismissSuccess(
+                            MyStoreViewModel.UTM_SOURCE,
+                            jitmId,
+                            featureClass
+                        )
+                    }
+                    else -> jitmTracker.trackJitmDismissFailure(
+                        MyStoreViewModel.UTM_SOURCE,
+                        jitmId,
+                        featureClass,
+                        response.error?.type,
+                        response.error?.message
+                    )
+                }
+            }
+        }
+    }
+}
+
+
